style(#3): Fixing lint for file
