+++ .eslintrc.json
@@ -52,7 +52,7 @@
                 "parameters": 1,
                 "body": 1
             },
-            "Function Expression": {
+            "FunctionExpression": {
                 "parameters": 1,
                 "body": 1
             },
